requirements.txt
requests
python-telegram-bot==20.8
flask
